pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nn-cz-sentiment"
version = "0.1.0"
description = "Czech sentiment analysis models and training utilities"
readme = "README.md"
authors = [{ name = "Adrian Vargas" }]
requires-python = ">=3.11,<4.0"
keywords = ["sentiment-analysis", "nlp", "czech", "transformers", "machine-learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

# Core dependencies from requirements.txt
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "tqdm>=4.65.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "joblib>=1.3.0",
    "pyyaml>=6.0",
    "emoji>=2.0.0",
    "datasets>=2.12.0",
    "evaluate>=0.4.0",
]

[project.optional-dependencies]
# Development and testing dependencies
dev = [
    "ipykernel>=6.25.0",
    "jupyter>=1.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.4.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

# CLI entry points - main commands for the package
[project.scripts]
nn-cz-train = "src.train:main"
nn-cz-predict = "src.predict:main"

[project.urls]
Homepage = "https://github.com/yourusername/nn-cz-sentiment-final"
Repository = "https://github.com/yourusername/nn-cz-sentiment-final"
Issues = "https://github.com/yourusername/nn-cz-sentiment-final/issues"

# Package discovery - tell setuptools to find packages in src/
[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

# Include non-Python files
[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml", "*.json", "*.txt"]

[tool.black]
line-length = 88
target-version = ["py311"]
skip-string-normalization = false
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | \.venv311
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
skip_glob = ["*.egg-info", ".venv*", "build", "dist"]

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [".git", "__pycache__", ".venv", ".venv311", "build", "dist", "*.egg-info"]

[tool.mypy]
python_version = "3.11"
files = ["src", "scripts"]
ignore_missing_imports = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_untyped_defs = false
check_untyped_defs = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --disable-warnings"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
